feat: make typescript workflow for dump file analysis
